+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.1.7]
 - now create_canister always sends args, Option<Vec<u8>>
+- you can specify extra controllers when creating canisters
 
 ## [0.1.6]
 - auth rewritten to be async, and to just use function names